+++ src/theme/components/featuredContent.ts
@@ -27,31 +27,39 @@ const FeaturedContent = {
     }
     return {
       bgColor: "ui.bg.default",
+      left: full ? "50%" : "auto",
+      marginLeft: full ? "-50vw" : "auto",
+      marginRight: full ? "-50vw" : "auto",
+      position: "relative",
+      right: full ? "50%" : "auto",
+      width: full ? "100vw" : "100%",
       _dark: {
         bgColor: "dark.ui.bg.default",
       },
-      width: full ? "100vw" : "100%",
-      left: full ? "50%" : "auto",
-      right: full ? "50%" : "auto",
-      position: "relative",
-      marginLeft: full ? "-50vw" : "auto",
-      marginRight: full ? "-50vw" : "auto",
       wrapper: {
         ...wrapperStyles,
-        minHeight: "320px",
-        maxWidth: full ? "1280px" : "100%",
-        display: "flex",
         alignItems: "stretch",
+        display: "flex",
         flexDirection: imageAtEnd
           ? { sm: "column-reverse", md: "row-reverse" }
           : { sm: "column", md: "row" },
+        maxWidth: full ? "1280px" : "100%",
+        minHeight: "320px",
+        paddingLeft: full ? { base: null, md: "s" } : null,
+        paddingRight: full ? { base: null, md: "s" } : null,
       },
       text: {
         display: "flex",
-        padding: "l",
         flex: 1,
         flexDirection: "column",
         justifyContent: "center",
+        padding: "l",
+        /** The `paddingLeft` attribute is used to adjust the spacing around the
+         * text when the image is positioned at the end. For aesthetic reasons,
+         * we opted to not adjust the spacing around the text when the image is
+         * positioned at the start.
+         * */
+        paddingStart: full && imageAtEnd ? { base: null, md: 0 } : null,
       },
       imgWrapper: {
         backgroundPosition: "center",